Changed splits property to integer.

+++ src/main/java/com/after_sunrise/cryptocurrency/cryptotrader/core/PropertyController.java
@@ -33,7 +33,7 @@ public interface PropertyController extends PropertyManager {
 
     void setTradingAversion(String site, String instrument, BigDecimal value);
 
-    void setTradingSplit(String site, String instrument, BigDecimal value);
+    void setTradingSplit(String site, String instrument, Integer value);
 
     void setTradingDuration(String site, String instrument, Duration value);
